Print whenever reading from stdin. (IREE-9063)
We have a few tools that read from stdin by default. It's confusing when there is no difference in output between a correctly formatted tool invocation just taking a while (running benchmarks or compiling a program) compared to an invocation with a typo that is spinning waiting for input.

Some of these changes may be overly spammy, and IREE-5931 has other (complementary) suggestions on how to make these tools less error-prone.

---

I considered logging in the implementation of `iree_stdin_read_contents` itself (base/internal/file_io), but figured that each callsite can customize the message this way.