[vulkan] Create command pools with CREATE_RESET_COMMAND_BUFFER_BIT (IREE-9404)
According to ARM Mali GPU doc
https://developer.arm.com/documentation/101897/0200/cpu-overheads/command-pools-for-vulkan:

> Command pools do not automatically recycle memory from deleted command
> buffers unless created with the RESET_COMMAND_BUFFER_BIT flag. Pools
> without this flag do not recycle their memory until the application
> resets the pool.

This helps to address a memory leaking issue in `iree-benchmark-module`
where we run many repetitions/iterations of the model with the same
Vulkan device. The command pool's lifetime is associated with the
device; so it ends up we never release the resources occupied by
previous command buffers until the whole program exit.

This is not an issue for other GPUs (e.g., Adreno). However,
we will need this flag eventually anyway, given that ML workloads
does not have a clearcut frame concept where we can
`vkResetCommandPool` at the frame boundary. So we need to track
individual command buffers and reset them separately.