Switch tests to runtime skipping if drivers are missing. (IREE-9429)
Progress on IREE-8761

These two tests use specific HAL drivers but aren't really testing the drivers themselves. Rather than exclude the tests at configure time by branching in the CMake file, we can always include the tests, but just skip their test cases if the desired HAL driver is not included in the build.

This does risk always skipping the tests (e.g. if there was a typo in the driver name), but I think the CMake simplification is worth it.