Remove unnecessary visibility (IREE-10031)
In general, tests shouldn't be visible. If someone's depending on your
test, it's usually a bad sign. There were a few other places where
things were unnecessarily marked as visible.

We're generally pretty loose about visibility and make things public
because it doesn't function very well outside of a monorepo, but tests
are usually an exception.