Remove tooling/cpu_features, switch to base/internal/cpu. (IREE-10303)
Fix a bug in the ARM64 path in base/internal/cpu, flagged by the
warnings pasted below. The issue was some missing parentheses, but I
decided to reword this as an if() statement to make it less tricky.
```
/usr/local/google/home/benoitjacob/iree/runtime/src/iree/base/internal/cpu.c:43:3: error: operator '?:' has lower precedence than '|'; '|' will be evaluated first [-Werror,-Wbitwise-conditional-parentheses]
  IREE_SET_IF_HWCAP(hwcap, IREE_HWCAP_ASIMDDP, out_fields[0],
  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/local/google/home/benoitjacob/iree/runtime/src/iree/base/internal/cpu.c:28:23: note: expanded from macro 'IREE_SET_IF_HWCAP'
                      ? field_bit
```

Change the e2e matmul tests YAML trace format to not have "+" prefixes
on the required target features. That was done out of lazyness as we had
the "+" in the LLVM cpu features argument, so the old
tooling/cpu_features accepted +-prefixed feature keys, but the new
base/internal/cpu does not.