Microkernel for `linalg_ext.pack`, take 2 (IREE-10969)
This is the re-landing of IREE-10927, which was rolled back in IREE-10951 due to
GCC breakage.

Plus:
- The fix for the GCC fix that caused it to be reverted. It
was caused by a difference betwen G++ and Clang++: G++ defines
`__STDC__`, Clang++ does not.
- Lit-tests. I had inadvertently
forgotten that. This actually caught an issue, I had defined the type of
`padding_value` as `i8` in the `vmvx.pack.i8i8` import and that was a
mistake as narrow integers are always `i32` at that level. The reason
why the e2e tests hadn't caught that is that they are only testing
`pack.i32i32`.