[vulkan] Add support for VK_EXT_subgroup_size_control (IREE-11406)
This commit plumbs through support for VK_EXT_subgroup_size_control in
Vulkan HAL driver. The extension is marked as optional and requested
when available. SPIR-V executable FlatBuffer gains a field to encode the
requested subgroup size from CodeGen, 0 meaning no requirement.

This is to be used in a later commit to enable selecting wave32 mode for
AMD RDNA architectures.