Overhaul CMake and Bazel CUDA support (IREE-11651)
On the CMake side:

* Reworked SDK download so that it won't partially clobber on version
changes, etc.
* Only downloads the SDK if not already downloaded.
* Nice errors if missing deps that will keep downloads from working.
* Localized all of the CUDA special stuff into
build_tools/third_party/cuda (out of top-level CMake and leaf code
directories).
* Switched CUDA target/driver to bazel_to_cmake (and normalized
everything between the two).

On the Bazel side:

* There is now a Bazel build for CUDA :)
* SDK detection logic is primitive but functional. If the env var
IREE_CUDA_TOOLKIT_ROOT is set, then the SDK will be symlinked in and
configured.
* If the SDK is not available, then the HAL target and driver will not
be built.
* CUDA support should now be built as part of all CI actions.

I also had to fix bazel_to_cmake in a couple of unexpected ways.

This should integrate relatively cleanly into the Google systems. In
order to stage it better, I made the new BUILD files BUILD.bazel so that
they are ignored on that side. Once we get it in, we can rename them and
apply the couple of rewrites that will be needed to redirect to
Google-side SDK targets. The only thing that should be needed to make
this import is to add a rewrite for `"@iree_cuda//:enabled"` to a
private bool_flag that is hard-coded to False (until ready to enable).
Example:

```
bool_flag(
    name = "enabled",
    build_setting_default = False,
)
```

Co-authored-by: Geoffrey Martin-Noble <gcmn@google.com>