Add a `query_tile_sizes` ukernel. (IREE-11640)
This is part 1/9 of IREE-11755: Finish the transition to
`--iree-flow-enable-data-tiling`.

This returns the tile size values to be used for the given tensor
encoding, passed as ukernel `flags`. This PR reworks a bit the existing
`vmvx.query_tile_sizes` op, to make it take such flags, because
otherwise we would need custom logic in `rewriteToCall` in the
`VMVXToVM` pass.

This is properly implemented on ARM64, returning the correct tile size
for baseline / `+dotprod` / `+i8mm`. The standard default `8x4x8` matmul
tile sizes are returned in other cases.