Adding IREE_FLAG_LIST utility for repeated string flags. (IREE-11806)
Currently only supports strings but we can extend it in the future with
some macro magic.

Many of the tools were using C++ to implement repeated flags for things
like `--function_input=` and now they can just use the common base C
APIs instead of needing extra C++ goo. Though passing giant inputs in
flags is never going to be great we now avoid the alloc+copy for each
flag that was previously required to do this.