Discard root tasks before leaf tasks. (IREE-11853)
Fixes IREE-11843. The assertion there is saying that we should not be
discarding a task before its dependencies. From the comments at
https://github.com/iree-org/iree/blob/d91712bf4d32c44f6075620293e638418a47d78f/runtime/src/iree/hal/drivers/local_task/task_command_buffer.c#L51-L61
I understand that root tasks may be dependencies of leaf tasks, so it
makes sense to order the `iree_task_list_discard` calls accordingly?
This fixes the assertion observed in IREE-11843, anyway.