Centralize CPU features, fix `iree_cpu_lookup_data_by_key`, add a `iree-cpuinfo` tool (IREE-12498)
The starting point of this PR was that I realized that I had forgotten
to update `iree_cpu_lookup_data_by_key`.

That sent me looking for a way to have it automatically sync'd with the
rest of the code, resulting in the present `.inl` approach.

As a bonus, I added a `iree-cpuinfo` tool, also automatically sync'd
with the `.inl` and just a few lines of impl. This is a quick debugging
tool allowing to answer quickly the question "which CPU features are we
actually detecting on this machine". Obviously the point is not provide
the best CPU feature detection tool, the point is that it's by
construction bug-for-bug equivalent to IREE's actual CPU feature
detection.