Making iree-benchmark-trace work with stateful traces. (IREE-12552)
This reworks it from precaching calls and running over those to instead
reprocessing the entire trace YAML each iteration but only timing the
calls. This allows for pretty much all trace features (with a few
tweaks). `<stdin>` modules are supported by passing `--capture_stdin` to
the tool so that it captures it once and reuses it for all reads. Inputs
are reparsed each iteration so that traces which take ownership of the
buffers or mutate them in place don't interfere. By default devices and
modules are reused across iterations but that can be disabled with
`--reuse_devices=false` and `--reuse_modules=false`.

This approach does result in slower overall runs depending on trace
content as libyaml is stupid slow but the timings aren't impacted thanks
to tight scoping. In the future we could have a cached representation of
the traces that keep the buffers around and avoid the need for document
reparsing but I'd rather do that once the workflows around this have
stabilized. I fully expect this tool to be rewritten at some point but
at least now we can run stateful traces and benchmark things like the
caching allocator in real(ish) execution sequences.

Traces look like this, with a special "Timing Active" plot showing when
we're actually timing (during calls):

![image](https://user-images.githubusercontent.com/75337/223591555-a0f9a592-f1f3-40d1-b22f-03affc1b3e97.png)
Zooming in each call event can be seen getting timed:

![image](https://user-images.githubusercontent.com/75337/223591703-5744d63b-a4c4-4ea3-a07a-6bbca373caf5.png)