Rename all Bazel BUILD files to BUILD.bazel (IREE-12663)
These were historically just called "BUILD" with no file extension
because that's the name that Google's internal version of Bazel (Blaze)
uses. We've been inconsistently adding the extension and this cleans it
all up. In a world with multiple build systems, having a file extension
makes sense :-) For instance, it avoids conflicts with directories
named "build" (the required name in some processes) on case-insensitive
filesystems.

Accomplished with:

```shell
find . -type d \
\( -path './bazel-*' -o -path ./third_party -o -path './*.venv' \) \
-prune -false -o -name BUILD -print | xargs -I {} mv {} {}.bazel
```