Adding FatELF support to the embedded ELF loader. (IREE-12624)
This is implemented based on the textual spec from the FatELF project.
FatELF is dead and is unlikely to ever be merged into Linux but it
exists and is sufficient for our purposes of deploying plugins in
portable ELF formats. No one ever specified a file extension as it was
assumed it'd become part of .so but to avoid confusion I've used '.sos'.

A simple `iree-fatelf` tool is added for working with the FatELF files
and does not need anything from the original FatELF project. We can ship
this in releases to allow end-users building plugins to package up their
ELFs in a compatible format.

The FatELF format is documented here: https://icculus.org/fatelf/ +
https://github.com/icculus/fatelf