Adding support for loading VM modules from dynamic libraries. (IREE-13112)
Users can now produce native shared libraries that export one or more
symbols used to create VM modules allowing for multiple different
modules to be packaged in the same binary. These dynamic modules receive
a key-value list of parameters that they can use to customize their
behavior that can be programmatically specified by whatever is loading
them (python kwargs, or flag `?key=value&key=value` on the command
line).

On the hosting runtime side `iree_vm_dynamic_module_load_from_file`
makes it easy to load and use the modules as if they were local to the
binary. The command line tools have been updated to support multiple
`--module=` flags and load each module into the context. This also
allows for multiple vmfbs to be loaded and all kinds of linkage behavior
(could have a training driver in a native module that calls functions in
a compiled module loaded earlier, etc).

As part of this multiple `--module=` flags can be provided now to tools
for both bytecode and dynamic modules. To ensure all the tools are using
the same flag helper/loader I've switched iree-benchmark-module,
iree-check-module, and iree-run-module to require explicitly specifying
`--module=-` to load the module from stdin. This removes the need for
the "reading from stdin message" that uglied up stdout.

I wasn't able to figure out Tracy support here (at least on Windows) as
if tracing is globally enabled the shared libraries will get Tracy
linked in and start a new tracing context. It's noted today and a
warning is emitted when the sample is compiled with tracing enabled.