ukernels: let `pack` take `padding_value` by value (IREE-13233)
It was a void* pointer to achieve type-genericity, but passing by value
is more friendly to compilers generating calls to this.

Some care went into not baking in endianness here, particularly in the
interface exposed to the compiler, and also within the ukernel
implementation. In cases where the element type is smaller than the
padding value type, we specify which part of the padding value is used
purely in terms of bit-significance and not in terms of byte-address. In
the implementation this is implemented with value-casting the padding
value to narrower types, instead of memcpy of bytes from the padding
value