ukernels: pack: move the generation of i64 padding_value to codegen (IREE-13264)
IREE-13233 changed the `pack` ukernel to take the `padding_value` as a i64.
Something has to convert from the `tensor.pack` `padding_value` (in the
element type) to i64. IREE-13233 let that be done by the vmvx/module.c entry
point. In order to have codegen call ukernels without having to
duplicate VMVX logic, the present PR moves this to the `LowerToUKernels`
pass.