Rework iree-run-mlir to operate against the IREE compiler C API. (IREE-12715)
This removes some unused functionality while sprinkling in the seeds of
something better. Future changes can go deeper on inferring compiler
configuration from available devices and such (will need some reworking
of run_module.c - I erred on the side of simplicity/reuse today). This
isn't meant to be an example of the best way to build an online compiler
but instead serve as an in-tree test of something that approximates one.