Support mhlo.all_to_all with NCCL (IREE-13326)
Adds support for `mhlo.all_to_all`. Since NCCL only supports all-to-all
on dimension 0, transpose and reshape are added during MHLOToLinalg
lowering to support any value of `split_dimension` and
`concat_dimension`.

This implementation maintains the op as a new CollectiveKind::AllToAll
all the way down until the runtime, where it is finally mapped to
ncclSend and ncclRecv. Alternatively, instead of introducing a new
CollectiveKind we could've lowered it earlier to a set of Send/Recv ops.
I think all-to-all is ubiquitous enough to warrant its own
CollectiveKind, but please let me know if there is any concerns here.

Fixes IREE-13101