Adding collective channel splitting to flow/stream/hal. (IREE-13578)
This requires hoisting channel memoization to the frontend level as
that's the only place that can know exactly which channels are required
and their topology. This PR removes the existing memoization that was
happening in the stream dialect but that should be fine for our current
demos that are doing one-shot execution via iree-run-module as the
memoized channels were being used once. Because memoization is moving up
the `stream.channel.default` op has been removed and future PRs may also
change `flow.channel.default` around to be more like
`stream.channel.create`.

Follow-on PRs will port the MHLO channel group mappings in IREE-13089 to use
these new split ops and add a memoization pass to that pipeline.

Inspiration for this comes from @okkwon's IREE-13089 and the recent addition
of `ncclCommSplit` to NCCL that allows us to expose the
`MPI_Comm_split`-like API consistently across targets.

Co-authored-by: Okwan Kwon <okwan@google.com>