`iree_c_embed_data` improvements (IREE-13814)
* Reconcile CMake with Bazel by dropping the CMake-specific
`GENERATED_SRCS` argument and instead just putting everything in `SRCS`,
like in Bazel.
* To bridge that difference, `bazel_to_cmake_converter.py` was facing
the impossible task of choosing, for each `srcs` entry, whether to put
it in `SRCS` or `GENERATED_SRCS`. It was simply deciding that based on
whether the leading character was a `:`, offering a way to distinguish
generated files in the case where they were generated within the current
package. But that didn't generalize to generated files from other
packages.
* In `SRCS` support various syntaxes for files in the source or in the
build directory (generated files) and automatically do the right thing
to match Bazel. In particular, make it simple to reference generated
files in a subdirectory.
* Concretely, to address IREE-13804, ukernel bitcode will need to reference
generated bitcode in architecture-specific subdirectories.
* In `generate_embed_data_main.cc`, generate clear errors on failure to
open an input/output file (would've saved some head scratching).

For some reason, this exposed an issue in `hal/cts`. The filenames were
missing a `"${TARGET_BACKEND}_"` prefix and somehow that was still
working; now with my changes that wasn't working anymore. Fixed by
adding the missing prefixes.

(This commit is one of a chain in completing
IREE-13804).