Fixes and simplifications to CPU-data handling (IREE-13881)
In DispatchABI.cpp:
* Was not checking if CPU feature bits in the schema applied to the
target architecture.
* Was not checking if CPU feature bits belonged to field 0 even though
the current code only handles that. Other fields are currently unused
but at least now we will get an assert failure if that changes.
* Was building a vector of bits, then creating one `LLVM::OrOp`for each
bit. Instead, why not OR those bits in C++ and create a single
`LLVM::OrOp` with the resulting OR-ed value?
* Was not checking that the first character of each feature string was a
`'+'` before dropping it.
* Was room for simplification in how the schema was used.

Additionally, folding into this PR a simplification to `iree-cpuinfo` (a
standalone tool to report the host CPU's features exactly as detected by
IREE). It was quadratic-time, as it unnecessarily called
`iree_cpu_lookup_data_by_key` for each feature.