Resetting collective batch when the CUDA command buffer arena is set. (IREE-13978)
The stream command buffer in CUDA is special and reused for multiple
submissions. The arena is reset each time a new submission comes in but
the collective batch was potentially hanging on to a block of memory the
arena had used. This always refreshes the collective batch each time the
arena is reset.