Adding support for async memory pool allocations in the CUDA HAL. (IREE-13440)
These aren't actually async as the CUDA HAL is synchronous but will make
use of CUDA's memory pooling features to reduce the alloc/free cost in a
way more friendly to CUDA's memory management than just the normal
allocator.

With this our queue-ordered allocations in CUDA now average a few us
(for me) and the caching allocator (or any other) is only needed for
caching non-queue-ordered allocations. A few compiler tweaks to switch
all allocations to queue-ordered will mean only explicitly allocated
buffers (constants/variables, stuff the user does manually, etc) will
not route to the pools.

It'd also be possible to explore using the same pools the queue-ordered
allocations use for explicit synchronous allocations (at least the
device-local pool) but it'd be nicer to get those out of the critical
path and then keep the pools separate such that the transient pool isn't
filled with persistent allocations.

Due to IREE-13984 this relies on the `--iree-stream-emulate-memset` flag
added in IREE-13994 being set when graphs are enabled. Since this is not the
default path today and there's just two test suites using it we just
flip the flag for them.