Improved f16/bf16 <-> f32 software conversions. (IREE-14392)
As I get ready to write f16/bf16 microkernels that use actual f16/bf16
CPU instructions, I need to get the "software" conversions closer to the
behavior of the CPU instructions, particularly switching to rounding to
nearest-even instead of naively truncating (which rounds towards zero)
and consistently handling corner cases with NaN, Inf and denormals (the
latter we continue to flush to zero).

Benefits:
1. As this is now aligned with the behavior of Clang/GCC builtin
`_Float16` type, we get to use that conditionally on compiler support.
This moves towards clearly defining the purpose of the software impl as
a reference impl, that should be more readable and correct than
performant.
2. Aligning closer to real hw instructions means that when we have to
write fuzzy compares in tests, we will need lower tolerance thresholds.
This also means that software fallbacks (e.g. f16 microkernels when the
device CPU does not have f16 instructions) are closer in behavior to the
faster code paths, only differing in speed.
3. Doing this now prevents entrenching the current behavior of these
helpers in golden values in tests -- already this PR has to update some
golden values in a CUDA test.
4. Round-to-nearest-even is inherently better than round-towards-zero in
that it's unbiased. For tests involving accumulating many values, such
as matmuls with substantially large K dimension size, rounding towards
zero would result in a large e2e inaccuracy , about 0.1%, as all the
roundings add up.