[metal] Retain device until command buffer completion (IREE-15288)
We use resource sets in `iree_hal_metal_device_queue_execute` to keep
track of command buffers and semaphores and release them when the
command buffers complete. It would require the underlying block pool
backing the resource sets to have a larger lifetime to make sure we
don't access destroyed data structures. So we need to retain the device
until the command buffer completes.

Fixes IREE-14814