[rocm] Excises (almost) dependence on /opt/rocm from the compiler. (IREE-16803)
Almost all uses were for bitcode linking, which we have been bundling
with the compiler for a long time.

I left one ancillary use in the ukernel directory as that depends on
some HIP headers. We can sever that dependency too, but will do that in
a separate PR as it isn't related.

Also takes this chance to:

* Remove foot-guns around the use of `--iree-rocm-link-bc` flag by
defaulting it to true. Afaict, it was only false because in the early
phases it was hard to ensure that everyone had bitcode files available.
* Remove all CMake plumbing that was setting `--iree-rocm-link-bc` or
`--iree-rocm-bc-dir`, instead making the build system work with defaults
for tests.
* Revamped error handling logic in ROCMTarget, which was really quite
awful. It was spewing errors to stderr and just barreling on in some
situations. All failure messages have been moved to the point
encountered and handled properly with an MLIR Location.
* The above uncovered a bug where we were segfaulting on tests that were
missing an ordinal attribute. These now properly error, and I fixed the
tests.
* Deletes the ROCM bazel lit tests since Bazel lacks the infrastructure
to bundle bitcode files and will likely never get it. We prefer to have
the build exercise default flag flows which precludes these tests being
the same between these systems.

I can't say I'm completely satisfied with the code quality of the
ROCMTarget yet, and it is clear that this was assembled from a few
different sources. However, the sloppy error handling was masking real
crash opportunities and has now been fixed.

---------

Co-authored-by: Scott Todd <scott.todd0@gmail.com>