Executable library call hooks system, and a sample Linux/CPU event implementation (IREE-15803)
There are two parts to this PR:
* The `runtime/` changes bring a minimal weak-hook extension point
around executable library calls. It's behind a C-internal flag: to
enable it, configure with `cmake
-DCMAKE_C_FLAGS=-DIREE_HAL_EXECUTABLE_LIBRARY_CALL_HOOK .` . When
enabled, it links to weak symbols and calls them when they are defined.
The symbols are `iree_hal_executable_library_call_hook_{begin,end}`.
* The `experimental/` changes bring one such hook implementation,
building a shared library defining such
`iree_hal_executable_library_call_hook_{begin,end}` symbols, doing Linux
event counter queries, with a number of events defined, imported from
JSON tables in the Linux source tree. At the moment it only defines
AMD-specific events and generic events, but more could be added. See the
README.md file there for how it's used.

Context:
This is a clean-up of the tools we had developed last fall to study
Llama2 on Zen4 CPUs. It's the most meaningful use of event counters that
I've been able to make yet. The difficulty with these is always to find
a way to make an apples-to-apples comparison. What we found in that
Llama2 work last fall was that some critical dispatch function had a
fluctuating latency, as shown graphically in Tracy histograms, which
raised the question: if this dispatch function has sometimes higher and
sometimes lower latency, can we understand that so that we could perhaps
more often fall in the lower-latency case? The next step was to
correlate these fluctuating latencies with some specific event counts.
This is what this tool is about.