[runtime][cts] Add test waiting on a semaphore for finite time and fix Vulkan driver (IREE-17126)
Fix Vulkan semaphore wait timeout if the timeout has already
passed. Don't check for a timeout on the host side.
Let Vulkan handle the timeout.
If the timeout has passed and the semaphore value is big enough,
we want an OK status to be returned.

Disable semaphore_test for Metal.