Adding `--iree-opt-import-parameters=` and cleaning up export. (IREE-16828)
The new option allows for importing all or some parameters from a
parameter file into the compiler. One or more parameter files to one or
more scopes can be specified, just as with the
`--parameters=[scope=]file.*` flag in the runtime tool. If a scope is
provided only parameters with that scope will be imported. If a
parameter (optionally within a scope) is named explicitly it will be
imported. If a maximum size is specified all parameters <= that size
will be imported.

The export pass has been reworked to be the inverse of the import such
that it's possible to round-trip through export and import. The import
pass doesn't currently support all of the types we can export as
`SerializableAttrInterface` doesn't have a deserialize method (yet) and
we need special handling of packed types (i1, i4, etc). For now warnings
are emitted and the globals are left as parameter references to still
allow partial imports.

Flag renames:
* `--iree-opt-parameter-archive-export-file=` +
  `--iree-opt-parameter-archive-export-scope=` ->
    `--iree-opt-export-parameters=[scope=]file`
* `--iree-opt-minimum-parameter-export-size=` ->
    `--iree-opt-export-parameter-minimum-size=`
* `--iree-opt-splat-parameter-archive-export-file=` ->
    `--iree-opt-splat-parameters=`