Strided pack/unpack ukernels (IREE-17516)
The pack ukernel gains the ability to support inner-strided sources, and
the unpack ukernel gains the ability to support inner-strided
destinations.

Note that the pack and unpack ukernels both use a small temporary buffer
to decouple source vs destination layout aspects, limiting the
combinatorics. The inner-striding aspect here is local to the "2D" side,
orthogonal to the "4D" side which is what we have architecture specific
code paths for. The "2D" side was handled by generic loops with generic
memcpy/memset primitives. What changes here is that now we need a
"memcpy with stride on either src or dst".

This is implemented as scalar loops. I tried using AVX-512
gather/scatter instructions, but that turned out 2x slower than a scalar
loop on AMD Zen4 where they are micro-coded, and the Intel documentation
suggests a very high cost there too.

Fixes IREE-17507.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>