Fixing iree_vm_ref_wrap_retain. (IREE-17610)
This was incorrect in the case where the out_ref contained the ptr and
would result in a leaked reference. The only usage of this in the
codebase is the vm.const.ref.rodata bytecode op and we seem to have
gotten lucky that there's not been any case where the same rodata was
loaded into the same register multiple times with no intervening
reassignment.