Optimizing queries for optional VM functions. (IREE-17823)
Queries with IMPORT_OPTIONAL bypass reporting and full status creation
as a way to avoid generating a bunch of stack traces for cases where the
caller can gracefully handle the absence of an import. This change adds
EXPORT_OPTIONAL to provide the same functionality to callers that know
the export may not exist. This speeds up startup (when __init is omitted
on a module), shutdown (when __deinit is omitted on a module), and event
notification (__notify).