Retaining binding tables and plumbing indirect cmds in local-task. (IREE-17838)
The task system doesn't currently support resubmission of task graphs
(as scheduling the graphs happens in-place in the task graph structure)
so it is emulated as with a deferred command buffer as with other
targets. Async targets need to retain the binding table contents until
the submission completes (as the resource set in the command buffer
won't) and a helper was added to allow that to be done efficiently in
the future (today it's still slow).