Adding flag placeholders to semaphores/events. (IREE-18122)
Events also gain a queue affinity that can be used to indicate which
queues a particular event may be set/waited on. The flags are currently
unused but will allow us to specify behavior modes in the future (such
as making exportable semaphores opt-in).

Since semaphores and events aren't yet exposed in the compiler there
were no changes needed there. The compiler will not (in common usage)
ever return new semaphores so will not need exportable flags and other
things we'll potentially add - applications will be setting those via
the C API.

Progress on IREE-18121 (event queue affinity needed for
[VK_KHR_device_group](https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VK_KHR_device_group.html)).