Adding iree_hal_deferred_buffer_t helper. (IREE-18158)
This is effectively just a subspan buffer but with the ability to
commit/decommit the backing storage. It's got a lot of footguns but
should mostly be used very carefully by binding layers or HAL
implementations that manage them.

This will be used to implement async allocations on CPU and any other
backend that does not natively support virtual memory. The intended
usage is for `iree_hal_device_queue_alloca` to reserve a deferred buffer
and then enqueue a commit in the queue timeline while
`iree_hal_device_queue_dealloca` will enqueue a decommit.