Fixing issues found when enabling indirect command buffers. (IREE-18382)
This is in preparation for making
`--iree-hal-indirect-command-buffers=true` the default as part of
IREE-17875.

Most of the fixes required were related to analysis failures that
required stricter handling of duplicate call graph traversal and
care around when initializers are combined (which still needs
improvement but at a general level not related to this work). A TODO was
fixed for supporting `stream.cmd.call` in reusable command buffers by
passing binding table ordinals along with buffers to the
`stream.cmd.func` ops once lowered into the HAL dialect - there are no
users of this functionality today but now it won't be a special case.